Support HM spacegroup with space and unspecified default spacegroup
- HM spacegroup are enclosed within parentheses and can contain space between
them as per the given format details in -
http://hincklab.uthscsa.edu/html/soft_packs/msi_docs/insight980/formats980/ClassicFiles.html.
The code doesn't recognize such spacegroup as it strictly look for 8 fields
after tokenizing them using default delimiters as space, tab, newline.

- It is legal to have .car files with PBC=ON and no space-group defined
in the fifth line and just have PBC and a,b,c,alpha,beta,gamma parameters.
Such structure is considered to have lowest symmetry space-group as P1. This
means there are no symmetry operations in play and all atoms are explicit
within the unit cell. The .car files for H2O and polymers fall under this
category since they are amorphous but are periodic.